refactor: remove legacy sniffings
