+++ packages/shared-base/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@dimensiondev/mask-wallet-core": "0.1.0-20211013082857-eb62e5f",
     "@masknet/base": "workspace:^",
+    "@masknet/flags": "workspace:^",
     "@masknet/encryption": "workspace:^",
     "@masknet/public-api": "workspace:^",
     "@masknet/typed-message": "workspace:^",
