+++ packages/shared-base/src/Account/index.ts
@@ -2,3 +2,9 @@ export interface Account<ChainId> {
     account: string
     chainId: ChainId
 }
+
+export enum SignType {
+    Message = 'message',
+    TypedData = 'typedData',
+    Transaction = 'transaction',
+}
