+++ packages/shared-base/src/MIME/index.ts
@@ -1,0 +1,5 @@
+export enum MimeType {
+    JSON = 'application/json',
+    Binary = 'application/octet-stream',
+    PNG = 'image/png',
+}
