--- packages/shared-base/src/Site/type.ts
@@ -1,15 +1,0 @@
-export enum EnhanceableSite {
-    Localhost = 'localhost',
-    Twitter = 'twitter.com',
-    Facebook = 'facebook.com',
-    Minds = 'minds.com',
-    Instagram = 'instagram.com',
-    OpenSea = 'opensea.io',
-    Mirror = 'mirror.xyz',
-}
-
-export enum ExtensionSite {
-    Dashboard = 'dashboard',
-    Popup = 'popup',
-    PopupConnect = 'popup-connect',
-}
