+++ packages/shared-base/src/index.ts
@@ -1,4 +1,5 @@
 export * from '@masknet/base'
+
 export * from './Account/index.js'
 export * from './Wallet/index.js'
 export * from './Messages/Mask.js'
@@ -12,6 +13,7 @@ export * from './Post/type.js'
 export * from './Pure/index.js'
 export * from './Routes/index.js'
 export * from './Site/index.js'
+export * from './MIME/index.js'
 export * from './Subscriptions/index.js'
 export * from './Social/index.js'
 export * from './Pageable/index.js'
