+++ packages/shared-base/src/utils/index.ts
@@ -12,16 +12,4 @@ export * from './personas.js'
 export * from './resolve.js'
 export * from './subscription.js'
 
-export enum SignType {
-    Message = 'message',
-    TypedData = 'typedData',
-    Transaction = 'transaction',
-}
-
-export enum MimeType {
-    JSON = 'application/json',
-    Binary = 'application/octet-stream',
-    PNG = 'image/png',
-}
-
 export type PartialRequired<T, RequiredKeys extends keyof T> = Omit<T, RequiredKeys> & Pick<Required<T>, RequiredKeys>
