+++ packages/web3-providers/src/types/RSS3.ts
@@ -1,7 +1,7 @@
+import type RSS3 from 'rss3-next'
 import type { Pageable, MimeType } from '@masknet/shared-base'
 import type { HubOptions } from '@masknet/web3-shared-base'
 import type { ChainId } from '@masknet/web3-shared-evm'
-import type RSS3 from 'rss3-next'
 /**
  * Conform to the RFC3339 Datetime format.
  * @example "2022-08-23T07:15:00Z"
