+++ packages/web3-shared/base/tsconfig.json
@@ -6,5 +6,5 @@
     "tsBuildInfoFile": "./dist/src.tsbuildinfo"
   },
   "include": ["./src", "./src/**/*.json"],
-  "references": [{ "path": "../../shared-base" }, { "path": "../../injected-script/sdk" }]
+  "references": [{ "path": "../../flags" }, { "path": "../../shared-base" }, { "path": "../../injected-script/sdk" }]
 }
