+++ packages/web3-shared/evm/tsconfig.json
@@ -8,7 +8,6 @@
   "include": ["./src", "./src/**/*.json"],
   "references": [
     { "path": "../base" },
-    { "path": "../flags" },
     { "path": "../../shared-base" },
     { "path": "../../injected-script/sdk" },
     { "path": "../../polyfills" }
